pyproject.toml
[project]
name = "book-api"
version = "0.1.0"
description = "A modern async API for managing books using FastAPI and SQLModel"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]",
    "sqlmodel",
    "aiosqlite",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
